Tweak EINTR-ready versions of read and write to return number of bytes processed like it POSIX counterparts do.

+++ utils.h
@@ -35,8 +35,8 @@ struct inotify_event* create_inotify_event (int         wd,
                                             const char *name,
                                             int        *event_len);
 
-int safe_read  (int fd, void *data, size_t size);
-int safe_write (int fd, const void *data, size_t size);
+ssize_t safe_read   (int fd, void *data, size_t size);
+ssize_t safe_write  (int fd, const void *data, size_t size);
 
 int is_opened (int fd);
 int is_deleted (int fd);
